Updated git submodules
Project: mdsal master 31dec8edafbe0d300f10c907d3e6bc299cbe52f6

Use TypedSchemaNode

Now that the getType() method is available through a unified
interface use it to simplify type checks.

Change-Id: I9538c3677daf4ac4e3eac080220403bf2f75adb8
Signed-off-by: Robert Varga <rovarga@cisco.com>


Remove an unneeded type specifier

Optional.of() can be correctly inferred with Java 8,
remove the type argument.

Change-Id: Ie0826c03c964bb1bdfd7a070fa1bf80cef2ac3f0
Signed-off-by: Robert Varga <rovarga@cisco.com>


Remove unneeded cast

Since we have corrected the getSchema() method, there is no need
to perform a cast.

Change-Id: Ib3c9b2d7dc3a15f2cb9e042055178106f6b33cae
Signed-off-by: Robert Varga <rovarga@cisco.com>


Fix incorrect null check

Eliminate an Eclipse-generated warning, which actually shows a bug.

Calls to childNonNull() did not pass an actual instance, but rather
a 'obj != null' expression. That expression got evaluated to
a boolean, which got boxed. Hence childNonNull() could not observe
a null no matter what and we'd get a NPE on next use -- exactly where
Eclipse put a warning.

childNonNull() returns a correctly annotatad @Nonnull reference,
which can be used as a guard against this kind of mistake: if the
original code used that reference, compilation would fail,
as the checked object is a Boolean and does not have a get() method.

This patch eliminates the expression and uses the returned reference,
eliminating the possiblity of a NPE.

Change-Id: I706f077f90ea38d3a48125beeaf4711182c8b8d5
Signed-off-by: Robert Varga <rovarga@cisco.com>


Use direct type checks

With the removal of old parser and ExtendedType each type
definition implements proper subtype, hence we do not have
to traverse the type tree up to root.

Change-Id: Ib7a6ffa5f221278547e27e37b19c14a1a3c26d8a
Signed-off-by: Robert Varga <rovarga@cisco.com>