Updated git submodules
Project: yangtools stable/beryllium 107826f9cd33c98f7e021e7147f1806e25b9c074

Bug 6410: Fixed initialization of typedefs in rpc

In the constructor of RpcEffectiveStatementImpl we initialize substatements
of an rpc statment. However, in the for loop we are incorrectly trying to
find type substatements instead of typedef substatements.

Typedef substatements in RpcEffectiveStatementImpl are now initialized correctly.

This patch is a manual cherry-pick of the following patch:
git.opendaylight.org/gerrit/#/c/44036/

Change-Id: Ic9a02727ffb4b6e6c798360e2266dc99d77c79e7
Signed-off-by: Igor Foltin <ifoltin@cisco.com>