Updated git submodules
Project: netvirt master 9368323907a69e68fb88519acab67ec09062a96e

Merge "BUG-6455 - due to blueprint migration, the order in which the respective ConfigActivator.start() methods is called for Netvirt and Netvirt-Providers is now random. This could cause timing issues where dependencies are not yet loaded when modules like NetVirt starts up. We need to inject the dependency of OvsdbInventoryService for NetvirtProviders so that NetvirtProviders has that service loaded before it starts."

BUG-6455 - due to blueprint migration, the order in which the respective ConfigActivator.start() methods is called for Netvirt and Netvirt-Providers is now random.
This could cause timing issues where dependencies are not yet loaded when modules like NetVirt starts up.
We need to inject the dependency of OvsdbInventoryService for NetvirtProviders so that NetvirtProviders has that service loaded before it starts.

Change-Id: I431554dbd76c9cad71e6d49916a8eb9e4d456f71
Signed-off-by: Bertrand <bertrand.low@hcl.com>