Updated git submodules
Project: netvirt stable/boron c2fffa27f2745fa9dfd24c0f616963b6686ed039

Merge "Bug 6556 - Fix to send the first unsolicited RA message for a Router interface" into stable/boron

Bug 6556 - Fix to send the first unsolicited RA message for a Router interface

On creation of the router interface, the 1st unsolicited RA is sent before
the association of the subnets with the interface. So the RA packet was not
transmitted as no valid gateway address was available for the interface.
The unsolicited RA is tranmitted after network/subnet association.

Change-Id: Ifcf8e0dbe395070ff2a1731b9cf2c151ed024e9d
Signed-off-by: kalaiselvik <Kalaiselvi_K@Dell.com>