Updated git submodules
Project: netvirt stable/boron fd8fd45bf4116690e995ac971937ebfb7b6d1c82

Bug 6758: Remove inter-VPN link state even if error

When an inter-VPN link in error state is removed, the
inter-VPN link is removed from MDSAL but not the state.
The bug was due because, when the state is error, it has
not any attribute with the endpoint states.

Change-Id: Ic8324c6ed5a5b5b3cf9a65694aff7ff137f3ae40
Signed-off-by: Alvaro Gonzalez ARROYO <alvaro.gonzalez.arroyo@ericsson.com>