Updated git submodules
Project: netvirt master bfab122640dff3342a93bc0a48bdc2ae44da7244

Bug 6758: Remove inter-VPN link state even if error

When an inter-VPN link in error state is removed, the
inter-VPN link is removed from MDSAL but not the state.
The bug was due because, when the state is error, it has
not any attribute with the endpoint states.

Change-Id: Ic8324c6ed5a5b5b3cf9a65694aff7ff137f3ae40
Signed-off-by: Alvaro Gonzalez ARROYO <alvaro.gonzalez.arroyo@ericsson.com>