Updated git submodules
Project: netvirt stable/boron bb48a9e1a04c06a1fe7576f7dfb02a75f9374aeb

Merge "Fix wiring issue in openstack.sfc-translator-impl" into stable/boron

Fix wiring issue in openstack.sfc-translator-impl

The bundle is pulling in the rpc-service RenderedServicePathService.
But in order to use it, we should use the odl-sfc-provider feature,
providing the implementation of this rpc-service, and also, registering
it in the OSGi service regisry.

Using the odl-sfc-model is exposing the class RenderedServicePathService
but don't register any implementation for this service in the OSGi
service registry.

Change-Id: I8b583bae9cdff7804b204d63ee88765a24679441
Signed-off-by: Alexis de Talhouët <adetalhouet@inocybe.com>